# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Little Imp daemon API",
    "version": "1",
    "description": "Local-only REST and MCP API exposed by littleimpd."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3210"
    }
  ],
...
  "paths": {
...
    "/bookmarks": {
      "post": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Save a URL and enqueue the ingestion pipeline.",
        "operationId": "postBookmarks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing active bookmark returned idempotently",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResponse"
                }
              }
            }
          },
          "201": {
            "description": "Bookmark created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "URL already exists in trash or archive",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Invalid URL or missing url field",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "POST"
      },
      "get": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "List active or archived bookmarks with filters and pagination.",
        "operationId": "getBookmarks",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "description": "Filter by tag name",
            "schema": {
              "type": "string",
              "description": "Filter by tag name"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "description": "Filter by exact domain",
            "schema": {
              "type": "string",
              "description": "Filter by exact domain"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "description": "Filter by exact category ID; takes precedence over category",
            "schema": {
              "type": "string",
              "description": "Filter by exact category ID; takes precedence over category"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "Filter by category name",
            "schema": {
              "type": "string",
              "description": "Filter by category name"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "description": "Inclusive ISO date or date-time lower bound",
            "schema": {
              "type": "string",
              "description": "Inclusive ISO date or date-time lower bound"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "description": "Inclusive ISO date or date-time upper bound",
            "schema": {
              "type": "string",
              "description": "Inclusive ISO date or date-time upper bound"
            }
          },
          {
            "name": "read_later",
            "in": "query",
            "required": false,
            "description": "Filter by read-later state; accepts boolean strings or numeric flags",
            "schema": {
              "type": "string",
              "description": "Filter by read-later state; accepts boolean strings or numeric flags",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            }
          },
          {
            "name": "read_state",
            "in": "query",
            "required": false,
            "description": "Filter by read state",
            "schema": {
              "type": "string",
              "description": "Filter by read state",
              "enum": [
                "read",
                "unread"
              ]
            }
          },
          {
            "name": "is_pinned",
            "in": "query",
            "required": false,
            "description": "Filter by pinned/starred state; accepts boolean strings or numeric flags",
            "schema": {
              "type": "string",
              "description": "Filter by pinned/starred state; accepts boolean strings or numeric flags",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            }
          },
          {
            "name": "opened_count_min",
            "in": "query",
            "required": false,
            "description": "Filter to bookmarks opened at least this many times",
            "schema": {
              "type": "integer",
              "description": "Filter to bookmarks opened at least this many times",
              "minimum": 0
            }
          },
          {
            "name": "opened_count_max",
            "in": "query",
            "required": false,
            "description": "Filter to bookmarks opened no more than this many times",
            "schema": {
              "type": "integer",
              "description": "Filter to bookmarks opened no more than this many times",
              "minimum": 0
            }
          },
          {
            "name": "last_opened_from",
            "in": "query",
            "required": false,
            "description": "Inclusive ISO date or date-time lower bound for last opened time",
            "schema": {
              "type": "string",
              "description": "Inclusive ISO date or date-time lower bound for last opened time"
            }
          },
          {
            "name": "last_opened_to",
            "in": "query",
            "required": false,
            "description": "Inclusive ISO date or date-time upper bound for last opened time",
            "schema": {
              "type": "string",
              "description": "Inclusive ISO date or date-time upper bound for last opened time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of results to return",
            "schema": {
              "type": "integer",
              "description": "Maximum number of results to return",
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of results to skip",
            "schema": {
              "type": "integer",
              "description": "Number of results to skip",
              "minimum": 0
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "description": "When true, return archived bookmarks",
            "schema": {
              "type": "string",
              "description": "When true, return archived bookmarks",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookmark page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkListResponse"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "GET"
      }
    },
    "/bookmarks/{id}": {
      "get": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Get one bookmark with extracted content.",
        "operationId": "getBookmarksId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookmark detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bookmark not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "GET"
      },
      "put": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Patch bookmark fields, tags, archive state, read state, and notes.",
        "operationId": "putBookmarksId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated bookmark",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Bookmark not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Invalid patch field",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "PUT"
      },
      "delete": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Soft-delete a bookmark by moving it to trash.",
        "operationId": "deleteBookmarksId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bookmark moved to trash"
          },
          "404": {
            "description": "Bookmark not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "DELETE"
      }
    },
...
    "/bookmarks/{id}/open": {
      "post": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Record a user-triggered external open for a bookmark.",
        "operationId": "postBookmarksIdOpen",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated bookmark open metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bookmark not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "POST"
      }
    },
    "/bookmarks/{id}/restore": {
      "post": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Restore a trashed bookmark.",
        "operationId": "postBookmarksIdRestore",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Restored bookmark",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bookmark not found or not in trash",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Restore succeeded but bookmark could not be fetched",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "POST"
      }
    },
    "/bookmarks/{id}/permanent": {
      "delete": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "Permanently delete a trashed bookmark.",
        "operationId": "deleteBookmarksIdPermanent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Bookmark permanently deleted"
          },
          "404": {
            "description": "Bookmark not found or not in trash",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "DELETE"
      }
    },
...
    "/bookmarks/{id}/related": {
      "get": {
        "tags": [
          "Bookmarks"
        ],
        "summary": "List semantically related bookmarks.",
        "operationId": "getBookmarksIdRelated",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "id path parameter",
            "schema": {
              "type": "string",
              "description": "id path parameter"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum related bookmarks",
            "schema": {
              "type": "integer",
              "description": "Maximum related bookmarks",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Related bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelatedBookmarksResponse"
                }
              }
            }
          },
          "404": {
            "description": "Bookmark not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Embedding provider is not configured",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-little-imp-source-method": "GET"
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807-style problem response",
        "properties": {
          "type": {
            "type": "string",
            "description": "Stable problem type URI"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable error title"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation",
            "nullable": true
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": false
      },
...
      "BookmarkResponse": {
        "type": "object",
        "description": "Single bookmark response",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Bookmark"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
      "BookmarkListResponse": {
        "type": "object",
        "description": "Paginated bookmark list",
        "properties": {
          "data": {
            "type": "array",
            "description": "Page items",
            "items": {
              "$ref": "#/components/schemas/Bookmark"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "additionalProperties": false
      },
...
      "BookmarkCreateRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "HTTP or HTTPS URL to save",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "description": "Optional title override"
          }
        },
        "required": [
          "url"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "localIntegrationBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Little Imp integration token",
        "description": "Managed local integration bearer token created with POST /integration-tokens."
      }
    }
  },
...
}